change variable type

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -7,7 +7,7 @@
 */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int l;
+	unsigned long int l;
 	int cont = 0;
 
 	l = (n ^ m);
